// src/types/ZendeskTicket.ts
export type ZendeskTicket = {
	allow_attachments: boolean;
	allow_channelback: boolean;
	assignee_email: string;
	assignee_id: number;
	attribute_value_ids: number[];
	brand_id: number;
	collaborator_ids: number[];
	collaborators: object[];
	comment: object;
	created_at: string;
	custom_fields: object[];
	description: string;
	due_at: string;
	email_cc_ids: number[];
	email_ccs: object;
	external_id: string;
	follower_ids: number[];
	followers: object;
	followup_ids: number[];
	forum_topic_id: number;
	group_id: number;
	has_incidents: boolean;
	id: number;
	is_public: boolean;
	macro_id: number;
	macro_ids: number[];
	metadata: object;
	organization_id: number;
	priority: string;
	problem_id: number;
	raw_subject: string;
	recipient: string;
	requester: object;
	requester_id: number;
	safe_update: boolean;
	satisfaction_rating: object;
	sharing_agreement_ids: number[];
	status: string;
	subject: string;
	submitter_id: number;
	tags: object[];
	ticket_form_id: number;
	type: string;
	updated_at: string;
	updated_stamp: string;
	url: string;
	via: object;
	via_followup_source_id: number;
	via_id: number;
	voice_comment: object;
};